pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "codesage"
version = "0.1.0"
description = "Production-grade CLI tool jo codebase analyze karke structured Markdown reports banata hai — accurate line-referenced issues ke saath."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Entropy Labs"}
]

dependencies = [
    "click>=8.1",
    "pydantic>=2.0",
    "pathspec>=0.12",
    "rich>=13.0",
    "questionary>=2.0",
    "openai>=1.0",
    "google-generativeai>=0.5",
    "anthropic>=0.25",
    "groq>=0.5",
    "ollama>=0.2",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[project.scripts]
codesage = "src.main:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
